relayer: move config validation to `ibc_relayer::config`
