+++ crates/relayer/Cargo.toml
@@ -69,6 +69,7 @@ secp256k1 = { version = "0.27.0", features = ["rand-std"] }
 strum = { version = "0.25", features = ["derive"] }
 tokio-stream = "0.1.14"
 once_cell = "1.17.1"
+tracing-subscriber = { version = "0.3.14", features = ["fmt", "env-filter", "json"] }
 
 [dependencies.byte-unit]
 version = "4.0.19"
@@ -100,11 +101,14 @@ features = ["rpc-client", "secp256k1", "unstable"]
 version = "0.34.0"
 default-features = false
 
+[dependencies.tendermint-light-client-verifier]
+version = "0.34.0"
+default-features = false
+
 [dev-dependencies]
 ibc-relayer-types = { version = "0.25.0", path = "../relayer-types", features = ["mocks"] }
 serial_test = "2.0.0"
 env_logger = "0.10.0"
-tracing-subscriber = { version = "0.3.14", features = ["fmt", "env-filter", "json"] }
 test-log = { version = "0.2.10", features = ["trace"] }
 
 # Needed for generating (synthetic) light blocks.
